refactor(server): extract scheduler route domain (#587)
Move the 10 scheduler routes (live/events/board/task_enable/disable/run/
task_events/start/stop/session_output) plus the scheduler-private helpers
_is_scheduler_running and _start_scheduler_daemon out of server.py into a
SchedulerRoutesMixin in agentwire/routes/scheduler.py, mirroring the
scratchpad/machines pattern. server.py composes the mixin and calls
register_scheduler_routes(self, self.app) in _setup_routes. Pure relocation —
route snapshot guard stays green. Also drops an orphaned scratchpad comment
left over from the scratchpad extraction.

Part of epic #560.

Closes #587

Built by dotdev.dev
